Minor Optimizations
Removed unnecessary functions that lingered as a result of my then lack of knowledge on how to make a program. Though, the impact on performance would be very minimal, it would hardly be noticeable.

Prior to this update, the program handled timing like this:
Event.Time -> Event.Seconds
Now, when the event is added:
Event.Seconds -> Event.Time
After the update, the added event now uses just the Event.Time that it already had to begin with, bypassing the need to convert and then reconvert.

Tested this new code a few times so it should work properly. Otherwise, I probably took out a load-bearing banana... oops!